restapi/requirements.txt
Django==4.1.5
django-debug-toolbar==3.8.1
djangorestframework==3.14.0
requests==2.26.0
drf_yasg==1.21.4
django_redis==5.0.0
psycopg2-binary==2.9.1
JSON-log-formatter==0.5.1
